pyproject.toml
[project]
name = "pcos"
version = "0.1.0"
description = "Project Closure OS CLI"
requires-python = ">=3.10"
dependencies = [
    "typer",
    "pyyaml",
    "rich",
    "requests",
    "python-dotenv",
    "google-api-python-client",
    "google-auth-oauthlib",
    "google-auth",
]

[project.scripts]
pcos = "pcos.cli:app"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
